# Conflicts:
#	LocalPackages/DataLayer/Package.swift

+++ LocalPackages/PresentationLayer/Sources/PresentationLayer/Pages/EntriesView/EntriesViewModel.swift
@@ -20,20 +20,26 @@
 //
 
 import Combine
+import DataLayer
 import Factory
 import Foundation
+import Macro
 import Models
+import SimpleToast
 
+// swiftlint:disable:next todo
+// TODO: remove ObservableObject when apple fixes the retain cycle of @State for @Observable
 @Observable
 @MainActor
-final class EntriesViewModel {
+final class EntriesViewModel: ObservableObject {
     var entries: [EntryUiModel] {
         guard !lastestQuery.isEmpty else {
             return (qaService.showMockEntries ? qaService.dataState.data : entryDataService.dataState.data) ?? []
         }
 
         let newResults = entryDataService.dataState.data?.filter {
-            $0.entry.name.lowercased().contains(lastestQuery)
+            $0.entry.name.lowercased().contains(lastestQuery) ||
+                $0.entry.issuer.lowercased().contains(lastestQuery)
         } ?? []
 
         return newResults
@@ -45,9 +51,13 @@ final class EntriesViewModel {
 
     var pauseCountDown = false
 
-    @ObservationIgnored var search = "" {
+    var focusSearchOnLaunch: Bool {
+        settingsService.focusSearchOnLaunch
+    }
+
+    @ObservationIgnored var query = "" {
         didSet {
-            searchTextStream.send(search)
+            searchTextStream.send(query)
         }
     }
 
@@ -55,9 +65,6 @@ final class EntriesViewModel {
 
     @ObservationIgnored
     private let searchTextStream: CurrentValueSubject<String, Never> = .init("")
-
-    @ObservationIgnored
-    private var pauseRefreshing = false
 
     @ObservationIgnored
     @LazyInjected(\ServiceContainer.qaService)
@@ -75,11 +82,17 @@ final class EntriesViewModel {
     @LazyInjected(\ServiceContainer.alertService)
     private var alertService
 
+    #if os(iOS)
+    @ObservationIgnored
+    @LazyInjected(\ToolsContainer.hapticsManager)
+    private var hapticsManager
+    #endif
+
     @ObservationIgnored
     @LazyInjected(\ServiceContainer.settingsService) private(set) var settingsService
 
     @ObservationIgnored
-    private var generateTokensTask: Task<Void, Never>?
+    @LazyInjected(\ServiceContainer.toastService) var toastService
 
     @ObservationIgnored
     private var cancellables = Set<AnyCancellable>()
@@ -108,7 +121,8 @@ final class EntriesViewModel {
         if offset < destination {
             targetIndex -= 1
         }
-        Task {
+        Task { [weak self] in
+            guard let self else { return }
             do {
                 try await entryDataService.reorderItem(from: offset, to: targetIndex)
             } catch {
@@ -127,6 +141,11 @@ extension EntriesViewModel {
         let code = entry.code.current
         assert(!code.isEmpty, "Code should not be empty")
         copyTextToClipboard(code)
+        toastService
+            .showToast(SimpleToast(title: #localized("Copied to clipboard", bundle: .module)))
+        #if os(iOS)
+        hapticsManager(.defaultImpact)
+        #endif
     }
 
     func toggleCodeRefresh(_ shouldPause: Bool) {
@@ -139,13 +158,26 @@ extension EntriesViewModel {
     }
 
     func delete(_ entry: EntryUiModel) {
-        Task {
-            do {
-                try await entryDataService.delete(entry)
-            } catch {
-                handle(error)
-            }
-        }
+        let action: ActionConfig = .init(title: "Yes",
+                                         titleBundle: .module,
+                                         role: .destructive,
+                                         action: { [weak self] in
+                                             guard let self else { return }
+                                             Task { [weak self] in
+                                                 guard let self else { return }
+                                                 do {
+                                                     try await entryDataService.delete(entry)
+                                                 } catch {
+                                                     handle(error)
+                                                 }
+                                             }
+                                         })
+        alertService.showAlert(.main(.init(title: "Delete entry",
+                                           titleBundle: .module,
+                                           // swiftlint:disable:next line_length
+                                           message: .localized("Are you sure you want to delete this entry? This action is irreversible.",
+                                                               .module),
+                                           actions: [action])))
     }
 }
